cdk/pyproject.toml
[project]
name = "workstation-cdk"
version = "0.1.0"
description = "AWS CDK infrastructure for EC2 workstation provisioning system"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib==2.152.0",
    "constructs>=10.0.0,<11.0.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.2",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
